+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## [Unreleased]
 
+- Upgraded Android SDK to Level 35
+  - Implemented keyboard edge-to-edge padding (bottom spacer)
+  - Opted out of Main Activity edge-to-edge
+- Fixed buggy key preview initialisation in landscape mode
 - Changed Help/About link colour to blue